+++ src/fideslang/validation.py
@@ -3,17 +3,16 @@ Contains all of the additional validation for the resource models.
 """
 
 import re
-from typing import List, Dict, Pattern
+from typing import Dict, List, Optional, Pattern, Set, Tuple
 
 from pydantic import ConstrainedStr
 
 from fideslang.default_fixtures import COUNTRY_CODES
 
-
 VALID_COUNTRY_CODES = [country["alpha3Code"] for country in COUNTRY_CODES]
 
 
-class FidesValidationError(Exception):
+class FidesValidationError(ValueError):
     """Custom exception for when the pydantic ValidationError can't be used."""
 
 
@@ -22,13 +21,15 @@ class FidesKey(ConstrainedStr):
     A FidesKey type that creates a custom constrained string.
     """
 
-    regex: Pattern[str] = re.compile(r"^[a-zA-Z0-9_.-]+$")
+    regex: Pattern[str] = re.compile(r"^[a-zA-Z0-9_.<>-]+$")
 
     @classmethod  # This overrides the default method to throw the custom FidesValidationError
     def validate(cls, value: str) -> str:
+        """Throws ValueError if val is not a valid FidesKey"""
+
         if not cls.regex.match(value):
             raise FidesValidationError(
-                f"FidesKeys must only contain alphanumeric characters, '.', '_' or '-'. Value provided: {value}"
+                f"FidesKeys must only contain alphanumeric characters, '.', '_', '<', '>' or '-'. Value provided: {value}"
             )
 
         return value
@@ -73,7 +74,7 @@ def matching_parent_key(value: FidesKey, values: Dict) -> FidesKey:
     parent_key_from_fides_key = ".".join(split_fides_key[:-1])
     if parent_key_from_fides_key != value:
         raise FidesValidationError(
-            "The parent_key ({0}) does not match the parent parsed ({1}) from the fides_key ({2})!".format(
+            "The parent_key ({0}) does match the parent parsed ({1}) from the fides_key ({2})!".format(
                 value, parent_key_from_fides_key, fides_key
             )
         )
@@ -93,3 +94,46 @@ def check_valid_country_code(country_code_list: List) -> List:
                     )
                 )
     return country_code_list
+
+
+def parse_data_type_string(type_string: Optional[str]) -> Tuple[Optional[str], bool]:
+    """Parse the data type string. Arrays are expressed in the form 'type[]'.
+
+    e.g.
+    - 'string' -> ('string', false)
+    - 'string[]' -> ('string', true)
+
+    These data_types are for use in DatasetField.fides_meta.
+    """
+    if not type_string:
+        return None, False
+    idx = type_string.find("[]")
+    if idx == -1:
+        return type_string, False
+    return type_string[:idx], True
+
+
+# Data types that Fides is currently configured to handle
+DATA_TYPE_NAMES: Set[str] = {
+    "string",
+    "integer",
+    "float",
+    "boolean",
+    "object_id",
+    "object",
+}
+
+
+def is_valid_data_type(type_name: str) -> bool:
+    """Is this type a valid data type identifier in fides?"""
+    return type_name is None or type_name in DATA_TYPE_NAMES
+
+
+def valid_data_type(data_type_str: Optional[str]) -> Optional[str]:
+    """If the data_type is provided ensure that it is a member of DataType."""
+
+    parsed_date_type, _ = parse_data_type_string(data_type_str)
+    if not is_valid_data_type(parsed_date_type):  # type: ignore
+        raise ValueError(f"The data type {data_type_str} is not supported.")
+
+    return data_type_str